requirements.txt
Package             Version
------------------- --------
asgiref             3.5.2
autopep8            1.7.0
Django              4.1.1
djangorestframework 3.14.0
pip                 22.0.2
pycodestyle         2.9.1
pytz                2022.2.1
setuptools          59.6.0
sqlparse            0.4.3
toml                0.10.2
